[node] tidy up and include test for processPath
